requirements.txt
pyjwt
pybranca
paseto